+++ src/tacty/opencv/preprocessing_pipeline.py
@@ -11,14 +11,32 @@ class PreProcessingPipeline:
 
     cleanImage: MatLike | None = None
     cleanImageNr: int | None = None
+    cleanImageDenoise: str | None = None
 
     def __init__(self, options: PreProcessingOptions, video: str):
         self.options = options
         self.video = video
 
+    def getDenoiseString(self) -> str:
+        return f"{self.options.denoiseEnabled}_{self.options.denoiseFilter}_{self.options.denoiseSize}"
+
+    def denoise(self, img: MatLike) -> MatLike:
+        size = self.options.denoiseSize
+        if self.options.denoiseFilter == "box":
+            return cv2.blur(img, (size, size))
+        elif self.options.denoiseFilter == "gaussian":
+            return cv2.GaussianBlur(img, (size, size), 0)
+        elif self.options.denoiseFilter == "median":
+            return cv2.medianBlur(img, size)
+        elif self.options.denoiseFilter == "bilateral":
+            return cv2.bilateralFilter(img, size, size*2, size/2)
+        else:
+            print(f"Unknown denoise filter: {self.options.denoiseFilter}.")
+            return img
+
     def removeBackground(self, img: MatLike) -> MatLike:
         # if we don't have the right frame already, grab it from the video
-        if(self.cleanImage is None or self.cleanImageNr != self.options.bgrFrame):
+        if(self.cleanImage is None or self.cleanImageNr != self.options.bgrFrame or self.cleanImageDenoise != self.getDenoiseString()):
             # open the video
             cap = cv2.VideoCapture(self.video)
 
@@ -28,9 +46,13 @@ class PreProcessingPipeline:
             if not success:
                 print(f"Failed to read frame {self.options.bgrFrame} from video {self.video}")
                 return img
+            
+            if self.options.denoiseEnabled:
+                frame = self.denoise(frame)
+
             self.cleanImage = frame
             self.cleanImageNr = self.options.bgrFrame
-
+            self.cleanImageDenoise = self.getDenoiseString()
             # close the video
             cap.release()
 
@@ -44,15 +66,18 @@ class PreProcessingPipeline:
         _, thresh = cv2.threshold(gray, int(self.options.bgrThreshold * 255), 255, cv2.THRESH_BINARY)
 
         # a bit of morphology to clean up the image
-        kernel = np.ones((3, 3), np.uint8)
-        thresh = cv2.morphologyEx(thresh, cv2.MORPH_OPEN, kernel)
-        thresh = cv2.morphologyEx(thresh, cv2.MORPH_CLOSE, kernel)
+        kernel_open = np.ones((5, 5), np.uint8)
+        kernel_close = np.ones((7, 7), np.uint8)
+        thresh = cv2.morphologyEx(thresh, cv2.MORPH_OPEN, kernel_open)
+        thresh = cv2.morphologyEx(thresh, cv2.MORPH_CLOSE, kernel_close)
 
         # apply the mask to the original image
         result = cv2.bitwise_and(img, img, mask=thresh)
         return result
     
     def process(self, img: MatLike) -> MatLike:
+        if self.options.denoiseEnabled:
+            img = self.denoise(img)
         if self.options.bgrEnabled:
             img = self.removeBackground(img)
         return img